added script to parse and sort answers
added python script to parse and sort the answers of ranking HIT. The
inputs are the HIT ID and the question number (which is either 1 or 2).
The log file (ranking_answer_log.csv) should be placed in the same
directory of this script. For example, run the script as:

python3 median_rank_parser.py 2J5JKM05BMJTOZKMPXKPDRA8V1CDIW 1