Removed keyword final for relation names
Removed the keyword final of the parameters which are relation names.
The relations such as INTERSECTION, UNION, FD, FD_PLUS should be
changable so that we can name different names to these tables for
different queries. Since generating these tables (esp. FD and FD+) is
expensive, we can store these tables in our database and reuse them in
the future, no redundant computation needed.