Issue #7 #10 -- Changed the condition
In IncremetalFD.java, we added tuples to corresponding tuple sets. On
line 221, we need to first determine if the tuple is already in the
tuple set. The function hasTuple used Tuple.isSame to do this. However,
this function also compares the source tables of these tuples (such as
Q1_RES, Q2_RES, etc). However, as long as these tuples are added into
the tuple set, there is no difference among them. So, it is
inappropriate to distinguish them through their source tables. What we
really care about is whether their values are equal. Therefore, this
commit added a new function isSameValue to do this job, only comparing
the values.