Fixed #16 -- Ignore if there is no foreign key.
This exception was caused by that a foreign key dependency table was not
found. However, if this table was not found, it means there was no
foreign key existing in the source table (in the FROM clause). So, when
a foreign key dependency table is not found, just ignore it. We don't
need to rewrite this kind of queries which have no foreign key
dependencis among their source tables.