Fixed #16 -- Clear the createdFKRelations each time
At the beginning of each round of a set of computation (intersection,
union, FD or/and FD+), we need to clear createdFKRelations which stores
the tables whose FK tables had been created  previously. In this way can
we make sure we have fresh and correct FK tables for different sets of
queries each time. In another word, the FK tables will be generated for
each round no matter the source tables are the same. The overhead is
very little and can be ignored, but it can ensure we have the correct
answers.

After this modification, the exception never came up again even though
we don't use a try-catch-finally to handle it.

After the examination of the rewritten queries for different sets of
queries, the results turn out to be correct.