Added or modified code for optimization.
Suppose @ is the FD predicate here.
Since the computation of FD can take a lot of time, we use the equation
R@S@T=(R@S)@T (suppose @ is the FD predicate) to reduce computation. The
basic idea is: if top(k-1) has been computed, then we can use the result
of top(k-1) and the new query to compute FD. This commit contains the
code for this purpose.