Optimized the computation again.
We can use previous FD table to ease our computation by just using the
new result table and the previous FD table as two inputs for the
computation of next FD table. Here, we can optimize it further. If the
new query is equivalent to a previous query which has already
contributed to generate previous FD table, then we can ignore this query
and just use the last FD table as this FD table.
The pseudo-code is:
CREATE TABLE THIS_FD AS SELECT * FROM PREVIOUS_FD